// src/generated/models/ContactsModel.ts
export const Contactsisautocreate = {
  0: 'No',
  1: 'Yes'
} as const;
export type Contactsisautocreate = keyof typeof Contactsisautocreate;
export const Contactsisbackofficecustomer = {
  0: 'No',
  1: 'Yes'
} as const;
export type Contactsisbackofficecustomer = keyof typeof Contactsisbackofficecustomer;
export const Contactsisprivate = {
  0: 'No',
  1: 'Yes'
} as const;
export type Contactsisprivate = keyof typeof Contactsisprivate;
export const Contactsleadsourcecode = {
  1: 'DefaultValue'
} as const;
export type Contactsleadsourcecode = keyof typeof Contactsleadsourcecode;
export const Contactsmarketingonly = {
  0: 'No',
  1: 'Yes'
} as const;
export type Contactsmarketingonly = keyof typeof Contactsmarketingonly;
export const Contactsmerged = {
  0: 'No',
  1: 'Yes'
} as const;
export type Contactsmerged = keyof typeof Contactsmerged;
// ...
export interface Contacts extends ContactsBase {
  accountidname?: string;
  accountidyominame?: string;
  accountrolecodename?: string;
  address1_addresstypecodename?: string;
  address1_composite?: string;
  address1_freighttermscodename?: string;
  address1_shippingmethodcodename?: string;
  address2_addresstypecodename?: string;
  address2_composite?: string;
  address2_freighttermscodename?: string;
  address2_shippingmethodcodename?: string;
  address3_addresstypecodename?: string;
  address3_composite?: string;
  address3_freighttermscodename?: string;
  address3_shippingmethodcodename?: string;
  adx_confirmremovepasswordname?: string;
  adx_identity_emailaddress1confirmedname?: string;
  adx_identity_locallogindisabledname?: string;
  adx_identity_lockoutenabledname?: string;
  adx_identity_logonenabledname?: string;
  adx_identity_mobilephoneconfirmedname?: string;
  adx_identity_twofactorenabledname?: string;
  adx_profilealertname?: string;
  adx_profileisanonymousname?: string;
  aging30?: string;
  aging30_base?: string;
  aging60?: string;
  aging60_base?: string;
  aging90?: string;
  aging90_base?: string;
  annualincome_base?: string;
  createdbyexternalpartyname?: string;
  createdbyexternalpartyyominame?: string;
  createdbyname?: string;
  createdbyyominame?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  creditlimit_base?: string;
  creditonholdname?: string;
  customersizecodename?: string;
  customertypecodename?: string;
  donotbulkemailname?: string;
  donotbulkpostalmailname?: string;
  donotemailname?: string;
  donotfaxname?: string;
  donotphonename?: string;
  donotpostalmailname?: string;
  donotsendmarketingmaterialname?: string;
  educationcodename?: string;
  entityimage_timestamp?: string;
  entityimage_url?: string;
  entityimageid?: string;
  exchangerate?: string;
  familystatuscodename?: string;
  followemailname?: string;
  fullname?: string;
  gendercodename?: string;
  haschildrencodename?: string;
  isautocreate?: Contactsisautocreate;
  isbackofficecustomername?: string;
  isprivate?: Contactsisprivate;
  isprivatename?: string;
  leadsourcecodename?: string;
  marketingonlyname?: string;
  mastercontactidname?: string;
  mastercontactidyominame?: string;
  merged?: Contactsmerged;
  mergedname?: string;
  modifiedbyexternalpartyname?: string;
  modifiedbyexternalpartyyominame?: string;
  modifiedbyname?: string;
  modifiedbyyominame?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  msa_managingpartneridname?: string;
  msa_managingpartneridyominame?: string;
  msdyn_disablewebtrackingname?: string;
  msdyn_isminorname?: string;
  msdyn_isminorwithparentalconsentname?: string;
  mspp_userpreferredlcidname?: string;
  onholdtime?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname?: string;
  parentcontactidname?: string;
  parentcontactidyominame?: string;
  parentcustomeridname: string;
  parentcustomeridyominame: string;
  participatesinworkflowname?: string;
  paymenttermscodename?: string;
  preferredappointmentdaycodename?: string;
  preferredappointmenttimecodename?: string;
  preferredcontactmethodcodename?: string;
  preferredsystemuseridname?: string;
  preferredsystemuseridyominame?: string;
  shippingmethodcodename?: string;
  slainvokedidname?: string;
  slaname?: string;
  statecodename?: string;
  statuscodename?: string;
  territorycodename?: string;
  timespentbymeonemailandmeetings?: string;
  transactioncurrencyidname?: string;
  versionnumber?: string;
  vs_ethnicitytypename?: string;
  vs_hearaboutusname?: string;
  yomifullname?: string;
  accountid?: object;
  _accountid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdbyexternalparty?: object;
  _createdbyexternalparty_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  masterid?: object;
  _masterid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedbyexternalparty?: object;
  _modifiedbyexternalparty_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  msa_managingpartnerid?: object;
  _msa_managingpartnerid_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  parentcontactid?: object;
  _parentcontactid_value?: string;
  preferredsystemuserid?: object;
  _preferredsystemuserid_value?: string;
  slaid?: object;
  _slaid_value?: string;
  slainvokedid?: object;
  _slainvokedid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}