// src/generated/models/AccountsModel.ts
export const Accountsaccountcategorycode = {
  1: 'PreferredCustomer',
  2: 'Standard'
} as const;
export type Accountsaccountcategorycode = keyof typeof Accountsaccountcategorycode;
export const Accountsaccountclassificationcode = {
  1: 'DefaultValue'
} as const;
export type Accountsaccountclassificationcode = keyof typeof Accountsaccountclassificationcode;
export const Accountsaccountratingcode = {
  1: 'DefaultValue'
} as const;
export type Accountsaccountratingcode = keyof typeof Accountsaccountratingcode;
export const Accountsaddress1_addresstypecode = {
  1: 'BillTo',
  2: 'ShipTo',
  3: 'Primary',
  4: 'Other'
} as const;
export type Accountsaddress1_addresstypecode = keyof typeof Accountsaddress1_addresstypecode;
export const Accountsaddress1_freighttermscode = {
  1: 'FOB',
  2: 'NoCharge'
} as const;
export type Accountsaddress1_freighttermscode = keyof typeof Accountsaddress1_freighttermscode;
export const Accountsaddress1_shippingmethodcode = {
  1: 'Airborne',
  2: 'DHL',
  3: 'FedEx',
  4: 'UPS',
  5: 'PostalMail',
  6: 'FullLoad',
  7: 'WillCall'
} as const;
export type Accountsaddress1_shippingmethodcode = keyof typeof Accountsaddress1_shippingmethodcode;
export const Accountsaddress2_addresstypecode = {
  1: 'DefaultValue'
} as const;
export type Accountsaddress2_addresstypecode = keyof typeof Accountsaddress2_addresstypecode;
export const Accountsaddress2_freighttermscode = {
  1: 'DefaultValue'
} as const;
export type Accountsaddress2_freighttermscode = keyof typeof Accountsaddress2_freighttermscode;
export const Accountsaddress2_shippingmethodcode = {
  1: 'DefaultValue'
} as const;
export type Accountsaddress2_shippingmethodcode = keyof typeof Accountsaddress2_shippingmethodcode;
export const Accountsbusinesstypecode = {
  1: 'DefaultValue'
} as const;
export type Accountsbusinesstypecode = keyof typeof Accountsbusinesstypecode;
export const Accountscreditonhold = {
  0: 'No',
  1: 'Yes'
} as const;
export type Accountscreditonhold = keyof typeof Accountscreditonhold;
export const Accountscustomersizecode = {
  1: 'DefaultValue'
} as const;
export type Accountscustomersizecode = keyof typeof Accountscustomersizecode;
export const Accountscustomertypecode = {
  1: 'Competitor',
  2: 'Consultant',
  3: 'Customer',
  4: 'Investor',
  5: 'Partner',
  6: 'Influencer',
  7: 'Press',
  8: 'Prospect',
  9: 'Reseller',
  10: 'Supplier',
  11: 'Vendor',
  12: 'Other'
} as const;
export type Accountscustomertypecode = keyof typeof Accountscustomertypecode;
export const Accountsdonotbulkemail = {
  0: 'Allow',
  1: 'DoNotAllow'
} as const;
export type Accountsdonotbulkemail = keyof typeof Accountsdonotbulkemail;
export const Accountsdonotbulkpostalmail = {
  0: 'No',
  1: 'Yes'
} as const;
export type Accountsdonotbulkpostalmail = keyof typeof Accountsdonotbulkpostalmail;
export const Accountsdonotemail = {
  0: 'Allow',
  1: 'DoNotAllow'
} as const;
export type Accountsdonotemail = keyof typeof Accountsdonotemail;
export const Accountsdonotfax = {
  0: 'Allow',
  1: 'DoNotAllow'
} as const;
export type Accountsdonotfax = keyof typeof Accountsdonotfax;
export const Accountsdonotphone = {
  0: 'Allow',
  1: 'DoNotAllow'
} as const;
export type Accountsdonotphone = keyof typeof Accountsdonotphone;
// ...
export const Accountsisprivate = {
  0: 'No',
  1: 'Yes'
} as const;
export type Accountsisprivate = keyof typeof Accountsisprivate;
export const Accountsmarketingonly = {
  0: 'No',
  1: 'Yes'
} as const;
export type Accountsmarketingonly = keyof typeof Accountsmarketingonly;
export const Accountsmerged = {
  0: 'No',
  1: 'Yes'
} as const;
export type Accountsmerged = keyof typeof Accountsmerged;
// ...
export interface AccountsBase {
  accountcategorycode?: Accountsaccountcategorycode;
  accountclassificationcode?: Accountsaccountclassificationcode;
  accountid: string;
  accountnumber?: string;
  accountratingcode?: Accountsaccountratingcode;
  address1_addressid?: string;
  address1_addresstypecode?: Accountsaddress1_addresstypecode;
  address1_city?: string;
  address1_country?: string;
  address1_county?: string;
  address1_fax?: string;
  address1_freighttermscode?: Accountsaddress1_freighttermscode;
  address1_latitude?: string;
  address1_line1?: string;
  address1_line2?: string;
  address1_line3?: string;
  address1_longitude?: string;
  address1_name?: string;
  address1_postalcode?: string;
  address1_postofficebox?: string;
  address1_primarycontactname?: string;
  address1_shippingmethodcode?: Accountsaddress1_shippingmethodcode;
  address1_stateorprovince?: string;
  address1_telephone1?: string;
  address1_telephone2?: string;
  address1_telephone3?: string;
  address1_upszone?: string;
  address1_utcoffset?: string;
  address2_addressid?: string;
  address2_addresstypecode?: Accountsaddress2_addresstypecode;
  address2_city?: string;
  address2_country?: string;
  address2_county?: string;
  address2_fax?: string;
  address2_freighttermscode?: Accountsaddress2_freighttermscode;
  address2_latitude?: string;
  address2_line1?: string;
  address2_line2?: string;
  address2_line3?: string;
  address2_longitude?: string;
  address2_name?: string;
  address2_postalcode?: string;
  address2_postofficebox?: string;
  address2_primarycontactname?: string;
  address2_shippingmethodcode?: Accountsaddress2_shippingmethodcode;
  address2_stateorprovince?: string;
  address2_telephone1?: string;
  address2_telephone2?: string;
  address2_telephone3?: string;
  address2_upszone?: string;
  address2_utcoffset?: string;
  adx_createdbyipaddress?: string;
  adx_createdbyusername?: string;
  adx_modifiedbyipaddress?: string;
  adx_modifiedbyusername?: string;
  businesstypecode?: Accountsbusinesstypecode;
  creditlimit?: string;
  creditonhold?: Accountscreditonhold;
  customersizecode?: Accountscustomersizecode;
  customertypecode?: Accountscustomertypecode;
  description?: string;
  donotbulkemail?: Accountsdonotbulkemail;
  donotbulkpostalmail?: Accountsdonotbulkpostalmail;
  donotemail?: Accountsdonotemail;
  donotfax?: Accountsdonotfax;
  donotphone?: Accountsdonotphone;
  donotpostalmail?: Accountsdonotpostalmail;
  donotsendmm?: Accountsdonotsendmm;
  emailaddress1?: string;
  emailaddress2?: string;
  emailaddress3?: string;
  entityimage?: string;
  fax?: string;
  followemail?: Accountsfollowemail;
  ftpsiteurl?: string;
  importsequencenumber?: string;
  industrycode?: Accountsindustrycode;
  lastonholdtime?: string;
  lastusedincampaign?: string;
  marketcap?: string;
  marketingonly?: Accountsmarketingonly;
  "msa_managingpartnerid@odata.bind"?: string;
  name: string;
  numberofemployees?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  ownershipcode?: Accountsownershipcode;
  "ParentAccountId@odata.bind"?: string;
  participatesinworkflow?: Accountsparticipatesinworkflow;
  paymenttermscode?: Accountspaymenttermscode;
  preferredappointmentdaycode?: Accountspreferredappointmentdaycode;
  preferredappointmenttimecode?: Accountspreferredappointmenttimecode;
  preferredcontactmethodcode?: Accountspreferredcontactmethodcode;
  "PreferredSystemUserId@odata.bind"?: string;
  "PrimaryContactId@odata.bind"?: string;
  primarysatoriid?: string;
  primarytwitterid?: string;
  processid?: string;
  revenue?: string;
  sharesoutstanding?: string;
  shippingmethodcode?: Accountsshippingmethodcode;
  sic?: string;
  "SLAId@odata.bind"?: string;
  stageid?: string;
  statecode: Accountsstatecode;
  statuscode?: Accountsstatuscode;
  stockexchange?: string;
  telephone1?: string;
  telephone2?: string;
  telephone3?: string;
  territorycode?: Accountsterritorycode;
  tickersymbol?: string;
  timezoneruleversionnumber?: string;
  "TransactionCurrencyId@odata.bind"?: string;
  traversedpath?: string;
  utcconversiontimezonecode?: string;
  vs_centrecode?: string;
  vs_primaryorganisationcode?: string;
  vs_referralcode?: string;
  "vs_State@odata.bind"?: string;
  "vs_wellbeingcentre@odata.bind"?: string;
  websiteurl?: string;
  yominame?: string;
}

export interface Accounts extends AccountsBase {
  accountcategorycodename?: string;
  accountclassificationcodename?: string;
  accountratingcodename?: string;
  address1_addresstypecodename?: string;
  address1_composite?: string;
  address1_freighttermscodename?: string;
  address1_shippingmethodcodename?: string;
  address2_addresstypecodename?: string;
  address2_composite?: string;
  address2_freighttermscodename?: string;
  address2_shippingmethodcodename?: string;
  aging30?: string;
  aging30_base?: string;
  aging60?: string;
  aging60_base?: string;
  aging90?: string;
  aging90_base?: string;
  businesstypecodename?: string;
  createdbyexternalpartyname?: string;
  createdbyexternalpartyyominame?: string;
  createdbyname?: string;
  createdbyyominame?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  creditlimit_base?: string;
  creditonholdname?: string;
  customersizecodename?: string;
  customertypecodename?: string;
  donotbulkemailname?: string;
  donotbulkpostalmailname?: string;
  donotemailname?: string;
  donotfaxname?: string;
  donotphonename?: string;
  donotpostalmailname?: string;
  donotsendmarketingmaterialname?: string;
  entityimage_timestamp?: string;
  entityimage_url?: string;
  entityimageid?: string;
  exchangerate?: string;
  followemailname?: string;
  industrycodename?: string;
  isprivate?: Accountsisprivate;
  isprivatename?: string;
  marketcap_base?: string;
  marketingonlyname?: string;
  masteraccountidname?: string;
  masteraccountidyominame?: string;
  merged?: Accountsmerged;
  mergedname?: string;
  modifiedbyexternalpartyname?: string;
  modifiedbyexternalpartyyominame?: string;
  modifiedbyname?: string;
  modifiedbyyominame?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  msa_managingpartneridname?: string;
  msa_managingpartneridyominame?: string;
  onholdtime?: string;
  owneridname: string;
  owneridyominame: string;
  ownershipcodename?: string;
  owningbusinessunitname?: string;
  parentaccountidname?: string;
  parentaccountidyominame?: string;
  participatesinworkflowname?: string;
  paymenttermscodename?: string;
  preferredappointmentdaycodename?: string;
  preferredappointmenttimecodename?: string;
  preferredcontactmethodcodename?: string;
  preferredsystemuseridname?: string;
  preferredsystemuseridyominame?: string;
  primarycontactidname?: string;
  primarycontactidyominame?: string;
  revenue_base?: string;
  shippingmethodcodename?: string;
  slainvokedidname?: string;
  slaname?: string;
  statecodename?: string;
  statuscodename?: string;
  territorycodename?: string;
  timespentbymeonemailandmeetings?: string;
  transactioncurrencyidname?: string;
  versionnumber?: string;
  vs_statename?: string;
  vs_wellbeingcentrename?: string;
  vs_wellbeingcentreyominame?: string;
  createdby?: object;
  _createdby_value?: string;
  createdbyexternalparty?: object;
  _createdbyexternalparty_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  masterid?: object;
  _masterid_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedbyexternalparty?: object;
  _modifiedbyexternalparty_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  msa_managingpartnerid?: object;
  _msa_managingpartnerid_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  parentaccountid?: object;
  _parentaccountid_value?: string;
  preferredsystemuserid?: object;
  _preferredsystemuserid_value?: string;
  primarycontactid?: object;
  _primarycontactid_value?: string;
  slaid?: object;
  _slaid_value?: string;
  slainvokedid?: object;
  _slainvokedid_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
  vs_state?: object;
  _vs_state_value?: string;
  vs_wellbeingcentre?: object;
  _vs_wellbeingcentre_value?: string;
}